// frontend-query-repository/src/app/models/datastream-port.ts
/* tslint:disable */
/* eslint-disable */
export interface DatastreamPort {

  /**
   * Name of connected Datastream to the Port
   */
  datastreamName?: string;
  description?: string;

  /**
   * internalip
   */
  internalip?: string;

  /**
   * ip
   */
  ip?: string;

  /**
   * port number
   */
  number?: number;
}